♻️ Refactor to using test model instead of nomal model for tests
