+++ tests/ModelCasterTest.php
@@ -1,12 +1,12 @@
 <?php
 
 use Morningtrain\DataTransferObjectCasters\Casters\ModelCaster;
-use Morningtrain\DataTransferObjectCasters\Models\TestModel;
+use Morningtrain\DataTransferObjectCasters\Tests\TestModel;
 use Spatie\DataTransferObject\Attributes\CastWith;
 use Spatie\DataTransferObject\DataTransferObject;
 
 it('casts to model from primary key', function () {
-    $model = TestModel::factory()->create(['name' => 'Clark']);
+    $model = TestModel::create(['name' => 'Clark']);
 
     $DTO = new class (['model' => $model->getKey()]) extends DataTransferObject {
         #[CastWith(ModelCaster::class, class: TestModel::class)]
@@ -19,7 +19,7 @@ it('casts to model from primary key', function () {
 });
 
 it('casts to model from specified column', function () {
-    $model = TestModel::factory()->create(['name' => 'Clark']);
+    $model = TestModel::create(['name' => 'Clark']);
 
     $DTO = new class (['model' => 'Clark']) extends DataTransferObject {
         #[CastWith(ModelCaster::class, class: TestModel::class, by: 'name')]
@@ -32,7 +32,7 @@ it('casts to model from specified column', function () {
 });
 
 it('casts to model value from specified column', function () {
-    $model = TestModel::factory()->create(['name' => 'Clark']);
+    $model = TestModel::create(['name' => 'Clark']);
 
     $DTO = new class (['modelName' => $model->getKey()]) extends DataTransferObject {
         #[CastWith(ModelCaster::class, class: TestModel::class, column: 'name')]
