+++ tests/TestCase.php
@@ -3,6 +3,7 @@
 namespace Morningtrain\DataTransferObjectCasters\Tests;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
+use Illuminate\Database\Schema\Blueprint;
 use Morningtrain\DataTransferObjectCasters\DataTransferObjectCastersServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
@@ -11,6 +12,8 @@ class TestCase extends Orchestra
     protected function setUp(): void
     {
         parent::setUp();
+
+        $this->setupDatabase();
 
         Factory::guessFactoryNamesUsing(
             fn (string $modelName) => 'Morningtrain\\DataTransferObjectCasters\\Database\\Factories\\'.class_basename($modelName).'Factory'
@@ -27,8 +30,14 @@ class TestCase extends Orchestra
     public function getEnvironmentSetUp($app)
     {
         config()->set('database.default', 'testing');
+    }
 
-        $migration = include __DIR__.'/../database/migrations/create_test_models_table.php.stub';
-        $migration->up();
+    private function setupDatabase(): void
+    {
+        $this->app['db']->connection()->getSchemaBuilder()->create('test_models', function (Blueprint $table) {
+            $table->id();
+            $table->string('name');
+            $table->timestamps();
+        });
     }
 }
